cpp: libraries: Add a C++ library to store data in shared memory
This library features:
- a C struct (shared_data_t) representing the shared data.
- a C++ class (SharedMemory) to create/attach/delete shared memory.
- a C++ class (WritePriorityLock) to create locks in shared memory,
with priority on write operations.
Locks are created for each element of the shared data structure,
identified by an enum.
WritePriorityLock also includes attributes and functions to emit
a signal when a share data element has been updated
and to wait for the signal in another process.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>