raspios: enable RTC overlay for PCF8523T
This commit adds the i2c-rtc overlay with the pcf8523 parameter to the
Raspberry Pi configuration. This enables the RTC functionality for the
PCF8523T chip, which is present on the RPi shield.

The Real-Time Clock (RTC) allows the system to maintain accurate
timekeeping, even when powered off, using the battery backup provided on
the shield.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>