tools: monitor: wizard: support choices grouped by category
The choices can now be grouped by category, each category will have
its own tab.

The choices list can now contain elements in the following format:
- value
- [value, category]
- [value, category, name]

If the category is not specified, it will default to "Other".
The name is used only for display. If not provided,
it will be set to str(value).
If there if only one category, the tab bar will be hidden.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>