tools: all: add sio_event attribute in config messages
Config messages are sent to the Monitor to open a modal to configure some
properties.
When a property is updated, a sio message is emitted with the new value.
To allow the receiver to have different handling methods depending on the
property, add an attribute "sio_event" to the config message to specify
the sio event handler that will receive the updated property.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>