cpp: libraries: fix ObstacleCircle nearest_point() method
As the bounding box margin is now an offset instead of a coefficient,
the nearest point on the circle is computed very far, even outside the
table.
Ensure the projected point remains on the circle boundary.
Also handle the special case where the input point is at the circle
center.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>