tools: detector: add property to enable Lidar scan delay compensation
Lidar scan introduces a delay between to the record of Lidar data
and the computation obstacle positions used by the Planner to
compute avoidance path.
The positions are computed using the current pose, while Lidar data
were recorded at a previous current pose, the delay depends on the Lidar
properties.
This commit introduces the sensor_delay property to compute obstacle
positions using a previous current pose. This delay represents the index
of the current pose to get in the buffer history.
The delay in milliseconds between two poses in the buffer depends on
the delay between two current poses sent by the firmware.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>