tools: monitor: add a timer to trigger simulated obstacle detection
Using a timer to trigger the update of each sensor should allow a more
regular update of obstacles than relying on the framerate to automatically
update sensors.
However, the update delay is still not precise at all because of
the amount of computation done in the event loop to update the 3D display.
Using a discrete GPU (like NVidia GPU using nvidia driver) really improves
the timer precision.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>