tools: planner: add option to choose default robot start position
Up to now, the robot's start position was assigned automatically using
a round-robin logic based on the robot's ID.
This behavior originates from a time when a single Planner
was responsible for all robots.

However, each robot now has its own dedicated Planner instance,
making this logic outdated and no longer relevant.

This commit adds a CLI option and an environment variable to the Planner
to set the start position explicitly.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>