tools: planner: prevent final action to be trigger during/after alignment
Using the wizard, the countdown is started during the alignment action.
If the countdown reaches 0 before the game starts, it will trigger
the final action, which is not wanted at this moment.

This commit sets the countdown to a negative value when the alignment
starts, so no action can be triggered.

The countdown will be reset to the game duration when the game starts.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>