tools: planner/copilot: store avoidance path in shared memory
Now the avoidance subprocess stores the computed path in a shared memory
PoseOrderList structure.

A new lock "AvoidancePath" is used to protect the read/write sections and
to signal when the path has been updated.

The Copilot uses this lock to wait for a new path, read the first pose
and send it to the firmware as the new pose order.

The "pose_order" and "path" Socket.IO signals are not emitted anymore
by the planner, however the server uses the lock to send these signals
to the "dashboard" namespace on path update.
This allows Dashboard en Monitor to remain functional even when executed
remotely (without access to the shared memory).

As a result, the avoidance subprocess could be run as a completely
standalone process (cogip-avoidance for example).

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>