tools: planner: remove sio_receiver_task in favor of asyncio.create_task()
The purpose of sio_receiver_task and sio_receiver_queue was to schedule
the execution of coroutines in the event loop in the background of the
function calling them.
This is now a dirty workaround, using asyncio.create_task() does the exact
same thing but in a much simpler and more recommended way.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>