utils: cache fnv1a_hash results for parameter lookups
Move fnv1a to utils folder because this is not a model and can be used
everywhere.

Add @cache decorator to avoid recomputing hashes for repeated parameter
names.

Parameter names are typically accessed multiple times during firmware
communication and hashes never changes, so we can cache it during the whole
execution lifetime.